% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entitySearchR-package.R
\docType{package}
\name{entitySearchR}
\alias{entitySearchR}
\alias{entitySearchR-package}
\title{API wrappers for legal entity searches}
\source{
http://pamiec.pl/pa/form/60,Zaloga-SS-KL-Auschwitz.html
}
\description{
API wrappers for legal entity searches
}
\author{
Alex Bresler (abresler@asbcllc.com)
}
