% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_nazis}
\alias{get_nazis}
\title{Get Nazi data}
\usage{
get_nazis(parse_bios = TRUE, tidy_columns = FALSE, return_message = TRUE)
}
\arguments{
\item{parse_bios}{if \code{TRUE} parses the detailed biography pages}

\item{tidy_columns}{if \code{TRUE} runs the  \code{\link{tidy_counts}} function}

\item{return_message}{if \code{TRUE} return a message after parsing}
}
\value{
a \code{data_frame}
}
\description{
This function scrapes and structures
the list of Nazis who served at Auschwitz.
}
\examples{
\dontrun{
get_nazis(parse_bios = TRUE, tidy_columns = TRUE,
return_message = TRUE)
}
}
\references{
\href{http://pamiec.pl/pa/form/60,Zaloga-SS-KL-Auschwitz.html}{ZAŁOGA SS KL AUSCHWITZ}
}
