% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_nazis}
\alias{import_nazis}
\title{Import Nazi cached  data}
\usage{
import_nazis(table_name = "all", visualize_results = FALSE,
  self_contained = FALSE, title = NULL)
}
\arguments{
\item{table_name}{the name of the table to import options are \itemize{
\item \code{all}: all tidy in wide form
\item \code{names}: unique nazi names
\item \code{items}: long data frame of columns that contain multiple items
}}

\item{visualize_results}{if \code{TRUE} returns a visual trelliscope display of the results}

\item{self_contained}{if \code{TRUE} htmlwidget is self-contained}

\item{title}{title of the trelliscope}
}
\value{
a \code{data_frame}
}
\description{
Returns a data frame or a visual
trelliscope display of the specified table.
The table contains data scraped on February 1st, 2017
}
\examples{
\dontrun{
nazihuntR::import_nazis(table_name = 'all')
nazihuntR::import_nazis(table_name = 'items')
}
}
