% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nazihuntr-package.R
\docType{package}
\name{nazihunter}
\alias{nazihunter}
\alias{nazihunter-package}
\title{API wrapper for SS Auschwitz personnel records}
\source{
http://pamiec.pl/pa/form/60,Zaloga-SS-KL-Auschwitz.html
}
\description{
API wrapper for SS Auschwitz personnel records
}
\author{
Alex Bresler (abresler@asbcllc.com)
}
