% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidy_counts}
\alias{tidy_counts}
\title{Tidy count columns}
\usage{
tidy_counts(data, column_keys = c("nameActualPerson", "urlBiography"),
  bind_to_original_df = FALSE)
}
\arguments{
\item{data}{a \code{data_frame}}

\item{column_keys}{the unique column keys \code{nameActualPerson} and \code{urlBiography}
selected by default}

\item{bind_to_original_df}{if \code{TRUE} binds the results back the original data as a nested data frame}
}
\value{
a if \code{bind_to_original_df} is \code{TRUE} a nested data frame else
a data frame
}
\description{
This function finds columns that contain
count data [multiple items] and tidies the columns
into long form.
}
\examples{
\dontrun{
import_nazis(table_name = 'all', visualize_results = FALSE) \%>\%
tidy_counts(column_keys = c('nameActualPerson', 'urlBiography'),
bind_to_original_df = TRUE)
}
}
