% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{visualize_nazis}
\alias{visualize_nazis}
\title{Visualize Nazi trelliscope}
\usage{
visualize_nazis(data, self_contained = FALSE, title = NULL)
}
\arguments{
\item{data}{a \code{data_frame}}

\item{self_contained}{if \code{TRUE} htmlwidget is self-contained}

\item{title}{Title of the trelliscope}
}
\value{
an \code{htmlwidget} of an interactive
\code{trelliscope} display
}
\description{
This function returns a trelliscope display
of the nazi data
}
\examples{
\dontrun{
nazihuntR::import_nazis(table_name = 'all') \%>\%
visualize_nazis()
}
}
