% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidy_counts}
\alias{tidy_counts}
\title{tidy counts}
\usage{
tidy_counts(data, column_keys = c("nameActualPerson", "urlBiography"),
  bind_to_original_df = FALSE)
}
\arguments{
\item{bind_to_original_df}{}
}
\description{
tidy counts
}
